+++ ArcdpsLogManager/CHANGELOG.md
@@ -6,6 +6,7 @@ This is the full changelog of the arcdps Log Manager.
 
 #### New features
 - Added a dialog to compress all uncompressed logs (Data → Compress logs) (thanks, @DBPhoenix!)
+- Added support for Bladesworn, Vindicator, Catalyst, Specter, Mechanist, and Untamed.
 
 ## Changes
 - Significantly reduced time spent loading the cache and searching for logs (big thanks to @Denrage!, issue #135)